Renamed Int Int64

+++ packages/vm/wasmproc/bytes.go
@@ -12,7 +12,7 @@ func NewBytesDecoder(data []byte) *BytesDecoder {
 }
 
 func (d *BytesDecoder) Bytes() []byte {
-	size := d.Int()
+	size := d.Int64()
 	if len(d.data) < int(size) {
 		panic("Cannot decode bytes")
 	}
@@ -21,7 +21,7 @@ func (d *BytesDecoder) Bytes() []byte {
 	return value
 }
 
-func (d *BytesDecoder) Int() int64 {
+func (d *BytesDecoder) Int64() int64 {
 	// leb128 decoder
 	val := int64(0)
 	s := 0
@@ -58,7 +58,7 @@ func NewBytesEncoder() *BytesEncoder {
 }
 
 func (e *BytesEncoder) Bytes(value []byte) *BytesEncoder {
-	e.Int(int64(len(value)))
+	e.Int64(int64(len(value)))
 	e.data = append(e.data, value...)
 	return e
 }
@@ -67,7 +67,7 @@ func (e *BytesEncoder) Data() []byte {
 	return e.data
 }
 
-func (e *BytesEncoder) Int(value int64) *BytesEncoder {
+func (e *BytesEncoder) Int64(value int64) *BytesEncoder {
 	// leb128 encoder
 	for {
 		b := byte(value)
